+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Chainlink adapters
+
+
 ## [[v1.0.3]((https://github.com/flare-foundation/flare-smart-contracts-v2/releases/tag/v1.0.3)])] - 2026-02-23
 
 ### Fixed
 
-- Relay contract
+* Relay contract
 
 
 ## [[v1.0.2]((https://github.com/flare-foundation/flare-smart-contracts-v2/releases/tag/v1.0.2)])] - 2026-01-19